@@ -24,7 +24,7 @@
 		<PackageTags>ipfs peer-to-peer p2p distributed file-system web</PackageTags>
 		<PackAsTool>true</PackAsTool>
 		<RootNamespace>Ipfs.Cli</RootNamespace>
-		<TargetFramework>net6.0</TargetFramework>
+		<TargetFramework>net9.0</TargetFramework>
 		<TargetLatestRuntimePatch>true</TargetLatestRuntimePatch>
 		<Title>C# IPFS CLI</Title>
 		<ToolCommandName>csipfs</ToolCommandName>
@@ -44,10 +44,12 @@
 
 	<ItemGroup>
 		<PackageReference Include="Common.Logging" Version="3.4.1" />
+		<PackageReference Include="Google.Protobuf" Version="3.31.1" />
 		<PackageReference Include="Ipfs.Http.Client" Version="0.33.0" />
-		<PackageReference Include="McMaster.Extensions.CommandLineUtils" Version="3.1.0" />
-		<PackageReference Include="Microsoft.AspNetCore.App" />
-		<PackageReference Include="System.Threading.Tasks.Dataflow" Version="6.0.0" />
+		<PackageReference Include="McMaster.Extensions.CommandLineUtils" Version="4.1.1" />
+		<PackageReference Include="System.Net.Http" Version="4.3.4" />
+		<PackageReference Include="System.Text.RegularExpressions" Version="4.3.1" />
+		<PackageReference Include="System.Threading.Tasks.Dataflow" Version="9.0.6" />
 	</ItemGroup>
 
 	<ItemGroup>
