@@ -9,15 +9,18 @@
     <!--<Exclude>[xunit.*]*</Exclude>-->
     <IsPackable>false</IsPackable>
     <RootNamespace>Ipfs.Engine</RootNamespace>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="coverlet.msbuild" Version="3.1.2" PrivateAssets="all" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.3.2" PrivateAssets="all" />
-    <PackageReference Include="MSTest.TestAdapter" Version="2.2.8" PrivateAssets="all" />
-    <PackageReference Include="MSTest.TestFramework" Version="2.2.10" PrivateAssets="all" />
-    <PackageReference Include="SharpZipLib" Version="1.3.3" />
+    <PackageReference Include="coverlet.msbuild" Version="6.0.4" PrivateAssets="all" />
+    <PackageReference Include="Google.Protobuf" Version="3.31.1" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.14.1" PrivateAssets="all" />
+    <PackageReference Include="MSTest.TestAdapter" Version="3.9.2" PrivateAssets="all" />
+    <PackageReference Include="MSTest.TestFramework" Version="3.9.2" PrivateAssets="all" />
+    <PackageReference Include="SharpZipLib" Version="1.4.2" />
+    <PackageReference Include="System.Net.Http" Version="4.3.4" />
+    <PackageReference Include="System.Text.RegularExpressions" Version="4.3.1" />
   </ItemGroup>
 
   <ItemGroup>
