@@ -1,20 +1,15 @@
 ﻿using McMaster.Extensions.CommandLineUtils;
-using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Ipfs.Cli
+namespace Ipfs.Cli.Commands;
+
+[Command(Name = "daemon", Description = "Start a long running IPFS deamon")]
+internal class DaemonCommand : CommandBase // TODO
 {
-    [Command(Description = "Start a long running IPFS deamon")]
-    class DaemonCommand : CommandBase // TODO
-    {
-        Program Parent { get; set; }
+    private Program Parent { get; set; }
 
-        protected override Task<int> OnExecute(CommandLineApplication app)
-        {
-            Ipfs.Server.Program.Main(new string[0]);
-            return Task.FromResult(0);
-        }
+    protected override Task<int> OnExecute(CommandLineApplication app)
+    {
+        Server.Program.Main([]);
+        return Task.FromResult(0);
     }
 }
