@@ -1,75 +1,69 @@
-﻿using Ipfs.Engine;
-using McMaster.Extensions.CommandLineUtils;
-using System;
-using System.Collections.Generic;
+﻿using McMaster.Extensions.CommandLineUtils;
 using System.ComponentModel.DataAnnotations;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Ipfs.Cli
+namespace Ipfs.Cli.Commands;
+
+[Command(Name = "dht", Description = "Query the DHT for values or peers")]
+[Subcommand(typeof(DhtFindPeerCommand))]
+[Subcommand(typeof(DhtFindProvidersCommand))]
+internal class DhtCommand : CommandBase
 {
-    [Command(Description = "Query the DHT for values or peers")]
-    [Subcommand("findpeer", typeof(DhtFindPeerCommand))]
-    [Subcommand("findprovs", typeof(DhtFindProvidersCommand))]
-    class DhtCommand : CommandBase
-    {
-        public Program Parent { get; set; }
+    public Program Parent { get; set; }
 
-        protected override Task<int> OnExecute(CommandLineApplication app)
-        {
-            app.ShowHelp();
-            return Task.FromResult(0);
-        }
+    protected override Task<int> OnExecute(CommandLineApplication app)
+    {
+        app.ShowHelp();
+        return Task.FromResult(0);
     }
+}
 
-    [Command(Description = "Find the multiaddresses associated with the peer ID")]
-    class DhtFindPeerCommand : CommandBase
+[Command(Name = "findpeer", Description = "Find the multiaddresses associated with the peer ID")]
+internal class DhtFindPeerCommand : CommandBase
+{
+    [Argument(0, "peerid", "The IPFS peer ID")]
+    [Required]
+    public string PeerId { get; set; }
+
+    private DhtCommand Parent { get; set; }
+
+    protected override async Task<int> OnExecute(CommandLineApplication app)
     {
-        DhtCommand Parent { get; set; }
+        Program Program = Parent.Parent;
 
-        [Argument(0, "peerid", "The IPFS peer ID")]
-        [Required]
-        public string PeerId { get; set; }
-
-        protected override async Task<int> OnExecute(CommandLineApplication app)
+        Peer peer = await Program.CoreApi.Dht.FindPeerAsync(new MultiHash(PeerId));
+        return Program.Output(app, peer, (data, writer) =>
         {
-            var Program = Parent.Parent;
-
-            var peer = await Program.CoreApi.Dht.FindPeerAsync(new MultiHash(PeerId));
-            return Program.Output(app, peer, (data, writer) =>
+            foreach (MultiAddress a in peer.Addresses)
             {
-                foreach (var a in peer.Addresses)
-                {
-                    writer.WriteLine(a.ToString());
-                }
-            });
-        }
+                writer.WriteLine(a.ToString());
+            }
+        });
     }
+}
 
-    [Command(Description = "Find peers that can provide a specific value, given a key")]
-    class DhtFindProvidersCommand : CommandBase
+[Command(Name = "findprovs", Description = "Find peers that can provide a specific value, given a key")]
+internal class DhtFindProvidersCommand : CommandBase
+{
+    [Argument(0, "key", "The multihash key or a CID")]
+    [Required]
+    public string Key { get; set; }
+
+    [Option("-n|--num-providers", Description = "The number of providers to find")]
+    public int Limit { get; set; } = 20;
+
+    private DhtCommand Parent { get; set; }
+
+    protected override async Task<int> OnExecute(CommandLineApplication app)
     {
-        DhtCommand Parent { get; set; }
+        Program Program = Parent.Parent;
 
-        [Argument(0, "key", "The multihash key or a CID")]
-        [Required]
-        public string Key { get; set; }
-
-        [Option("-n|--num-providers", Description = "The number of providers to find")]
-        public int Limit { get; set; } = 20;
-
-        protected override async Task<int> OnExecute(CommandLineApplication app)
+        IEnumerable<Peer> peers = await Program.CoreApi.Dht.FindProvidersAsync(Cid.Decode(Key), Limit);
+        return Program.Output(app, peers, (data, writer) =>
         {
-            var Program = Parent.Parent;
-
-            var peers = await Program.CoreApi.Dht.FindProvidersAsync(Cid.Decode(Key), Limit);
-            return Program.Output(app, peers, (data, writer) =>
+            foreach (Peer peer in peers)
             {
-                foreach (var peer in peers)
-                {
-                    writer.WriteLine(peer.Id.ToString());
-                }
-            });
-        }
+                writer.WriteLine(peer.Id.ToString());
+            }
+        });
     }
 }
