@@ -1,49 +1,43 @@
 ﻿using McMaster.Extensions.CommandLineUtils;
 using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Ipfs.Cli
+namespace Ipfs.Cli.Commands;
+
+[Command(Name = "id", Description = "Show info on an IPFS peer")]
+internal class IdCommand : CommandBase
 {
-    [Command(Description = "Show info on an IPFS peer")]
-    class IdCommand : CommandBase
+    [Argument(0, "peerid", "The IPFS peer ID")]
+    public string PeerId { get; set; }
+
+    private Program Parent { get; set; }
+
+    protected override async Task<int> OnExecute(CommandLineApplication app)
     {
-        [Argument(0, "peerid", "The IPFS peer ID")]
-        public string PeerId { get; set; }
+        MultiHash id = PeerId == null ? null : new MultiHash(PeerId);
+        Peer peer = await Parent.CoreApi.Generic.IdAsync(id);
+        using JsonWriter writer = new JsonTextWriter(app.Out);
+        writer.Formatting = Formatting.Indented;
 
-        Program Parent { get; set; }
-
-        protected override async Task<int> OnExecute(CommandLineApplication app)
+        writer.WriteStartObject();
+        writer.WritePropertyName("ID");
+        writer.WriteValue(peer.Id.ToBase58());
+        writer.WritePropertyName("PublicKey");
+        writer.WriteValue(peer.PublicKey);
+        writer.WritePropertyName("Adddresses");
+        writer.WriteStartArray();
+        foreach (MultiAddress a in peer.Addresses)
         {
-            MultiHash id = PeerId == null ? null : new MultiHash(PeerId);
-            var peer = await Parent.CoreApi.Generic.IdAsync(id);
-            using (JsonWriter writer = new JsonTextWriter(app.Out))
+            if (a != null)
             {
-                writer.Formatting = Formatting.Indented;
-
-                writer.WriteStartObject();
-                writer.WritePropertyName("ID");
-                writer.WriteValue(peer.Id.ToBase58());
-                writer.WritePropertyName("PublicKey");
-                writer.WriteValue(peer.PublicKey);
-                writer.WritePropertyName("Adddresses");
-                writer.WriteStartArray();
-                foreach (var a in peer.Addresses)
-                {
-                    if (a != null)
-                        writer.WriteValue(a.ToString());
-                }
-                writer.WriteEndArray();
-                writer.WritePropertyName("AgentVersion");
-                writer.WriteValue(peer.AgentVersion);
-                writer.WritePropertyName("ProtocolVersion");
-                writer.WriteValue(peer.ProtocolVersion);
-                writer.WriteEndObject();
+                writer.WriteValue(a.ToString());
             }
-            return 0;
         }
-
+        writer.WriteEndArray();
+        writer.WritePropertyName("AgentVersion");
+        writer.WriteValue(peer.AgentVersion);
+        writer.WritePropertyName("ProtocolVersion");
+        writer.WriteValue(peer.ProtocolVersion);
+        writer.WriteEndObject();
+        return 0;
     }
 }
