@@ -1,12 +1,8 @@
 ﻿using McMaster.Extensions.CommandLineUtils;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace Ipfs.Cli
+namespace Ipfs.Cli.Commands;
+
+[Command(Name = "pin", Description = "Manage data in local storage [WIP]")]
+internal class PinCommand : CommandBase // TODO
 {
-    [Command(Description = "Manage data in local storage [WIP]")]
-    class PinCommand : CommandBase // TODO
-    {
-    }
 }
