@@ -1,104 +1,98 @@
 ﻿using Ipfs.Engine;
 using McMaster.Extensions.CommandLineUtils;
-using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Ipfs.Cli
+namespace Ipfs.Cli.Commands;
+
+[Command(Name = "repo", Description = "Manage the IPFS repository")]
+[Subcommand(typeof(RepoGCCommand))]
+[Subcommand(typeof(RepoMigrateCommand))]
+[Subcommand(typeof(RepoStatCommand))]
+[Subcommand(typeof(RepoVerifyCommand))]
+[Subcommand(typeof(RepoVersionCommand))]
+internal class RepoCommand : CommandBase
 {
-    [Command(Description = "Manage the IPFS repository")]
-    [Subcommand("gc", typeof(RepoGCCommand))]
-    [Subcommand("migrate", typeof(RepoMigrateCommand))]
-    [Subcommand("stat", typeof(RepoStatCommand))]
-    [Subcommand("verify", typeof(RepoVerifyCommand))]
-    [Subcommand("version", typeof(RepoVersionCommand))]
-    class RepoCommand : CommandBase
+    public Program Parent { get; set; }
+
+    protected override Task<int> OnExecute(CommandLineApplication app)
     {
-        public Program Parent { get; set; }
-
-        protected override Task<int> OnExecute(CommandLineApplication app)
-        {
-            app.ShowHelp();
-            return Task.FromResult(0);
-        }
+        app.ShowHelp();
+        return Task.FromResult(0);
     }
+}
 
-    [Command(Description = "Perform a garbage collection sweep on the repo")]
-    class RepoGCCommand : CommandBase
+[Command(Name = "gc", Description = "Perform a garbage collection sweep on the repo")]
+internal class RepoGCCommand : CommandBase
+{
+    private RepoCommand Parent { get; set; }
+
+    protected override async Task<int> OnExecute(CommandLineApplication app)
     {
-        RepoCommand Parent { get; set; }
+        Program Program = Parent.Parent;
 
-        protected override async Task<int> OnExecute(CommandLineApplication app)
-        {
-            var Program = Parent.Parent;
-
-            await Program.CoreApi.BlockRepository.RemoveGarbageAsync();
-            return 0;
-        }
+        await Program.CoreApi.BlockRepository.RemoveGarbageAsync();
+        return 0;
     }
+}
 
-    [Command(Description = "Verify all blocks in repo are not corrupted")]
-    class RepoVerifyCommand : CommandBase
+[Command(Name = "stat", Description = "Repository information")]
+internal class RepoStatCommand : CommandBase
+{
+    private RepoCommand Parent { get; set; }
+
+    protected override async Task<int> OnExecute(CommandLineApplication app)
     {
-        RepoCommand Parent { get; set; }
+        Program Program = Parent.Parent;
 
-        protected override async Task<int> OnExecute(CommandLineApplication app)
-        {
-            var Program = Parent.Parent;
-
-            await Program.CoreApi.BlockRepository.VerifyAsync();
-            return 0;
-        }
+        CoreApi.RepositoryData stats = await Program.CoreApi.BlockRepository.StatisticsAsync();
+        return Program.Output(app, stats, null);
     }
+}
 
-    [Command(Description = "Repository information")]
-    class RepoStatCommand : CommandBase
+[Command(Name = "verify", Description = "Verify all blocks in repo are not corrupted")]
+internal class RepoVerifyCommand : CommandBase
+{
+    private RepoCommand Parent { get; set; }
+
+    protected override async Task<int> OnExecute(CommandLineApplication app)
     {
-        RepoCommand Parent { get; set; }
+        Program Program = Parent.Parent;
 
-        protected override async Task<int> OnExecute(CommandLineApplication app)
-        {
-            var Program = Parent.Parent;
-
-            var stats = await Program.CoreApi.BlockRepository.StatisticsAsync();
-            return Program.Output(app, stats, null);
-        }
+        await Program.CoreApi.BlockRepository.VerifyAsync();
+        return 0;
     }
+}
 
-    [Command(Description = "Repository version")]
-    class RepoVersionCommand : CommandBase
+[Command(Name = "version", Description = "Repository version")]
+internal class RepoVersionCommand : CommandBase
+{
+    private RepoCommand Parent { get; set; }
+
+    protected override async Task<int> OnExecute(CommandLineApplication app)
     {
-        RepoCommand Parent { get; set; }
+        Program Program = Parent.Parent;
 
-        protected override async Task<int> OnExecute(CommandLineApplication app)
-        {
-            var Program = Parent.Parent;
-
-            var stats = await Program.CoreApi.BlockRepository.VersionAsync();
-            return Program.Output(app, stats, null);
-        }
+        string stats = await Program.CoreApi.BlockRepository.VersionAsync();
+        return Program.Output(app, stats, null);
     }
+}
 
-    [Command(Description = "Migrate to the version")]
-    class RepoMigrateCommand : CommandBase
+[Command(Name = "migrate", Description = "Migrate to the version")]
+internal class RepoMigrateCommand : CommandBase
+{
+    [Argument(0, "version", "The version number of the repository")]
+    [Required]
+    public int Version { get; set; }
+
+    private RepoCommand Parent { get; set; }
+
+    protected override async Task<int> OnExecute(CommandLineApplication app)
     {
-        RepoCommand Parent { get; set; }
+        // TODO: Add option --pass
+        string passphrase = "this is not a secure pass phrase";
+        IpfsEngine ipfs = new(passphrase.ToCharArray());
 
-        [Argument(0, "version", "The version number of the repository")]
-        [Required]
-        public int Version { get; set; }
-
-        protected override async Task<int> OnExecute(CommandLineApplication app)
-        {
-            // TODO: Add option --pass
-            string passphrase = "this is not a secure pass phrase";
-            var ipfs = new IpfsEngine(passphrase.ToCharArray());
-
-            await ipfs.MigrationManager.MirgrateToVersionAsync(Version);
-            return 0;
-        }
+        await ipfs.MigrationManager.MirgrateToVersionAsync(Version);
+        return 0;
     }
-
 }
