@@ -1,30 +1,24 @@
 ﻿using McMaster.Extensions.CommandLineUtils;
-using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Ipfs.Cli
+namespace Ipfs.Cli.Commands;
+
+[Command(Name = "resolve", Description = "Resolve any type of name")]
+internal class ResolveCommand : CommandBase
 {
-    [Command(Description = "Resolve any type of name")]
-    class ResolveCommand : CommandBase
+    [Argument(0, "name", "The IPFS/IPNS/... name")]
+    [Required]
+    public string Name { get; set; }
+
+    [Option("-r|--recursive", Description = "Resolve until the result is an IPFS name")]
+    public bool Recursive { get; set; }
+
+    private Program Parent { get; set; }
+
+    protected override async Task<int> OnExecute(CommandLineApplication app)
     {
-        [Argument(0, "name", "The IPFS/IPNS/... name")]
-        [Required]
-        public string Name { get; set; }
-
-        [Option("-r|--recursive", Description = "Resolve until the result is an IPFS name")]
-        public bool Recursive { get; set; }
-
-        Program Parent { get; set; }
-
-        protected override async Task<int> OnExecute(CommandLineApplication app)
-        {
-            var result = await Parent.CoreApi.Generic.ResolveAsync(Name, Recursive);
-            app.Out.Write(result);
-            return 0;
-        }
-
+        string result = await Parent.CoreApi.Generic.ResolveAsync(Name, Recursive);
+        app.Out.Write(result);
+        return 0;
     }
 }
