@@ -1,12 +1,8 @@
 ﻿using McMaster.Extensions.CommandLineUtils;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace Ipfs.Cli
+namespace Ipfs.Cli.Commands;
+
+[Command(Name = "update", Description = "Download the latest version [WIP]")]
+internal class UpdateCommand : CommandBase // TODO
 {
-    [Command(Description = "Download the latest version [WIP]")]
-    class UpdateCommand : CommandBase // TODO
-    {
-    }
 }
