@@ -2,13 +2,13 @@
 
 /*
  * Copyright � 2002-2009 the original author or authors.
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -18,10 +18,6 @@
 
 #endregion
 
-using System;
-#if !SILVERLIGHT
-using System.Collections.Generic;
-#endif
 using System.Text;
 
 namespace Common.Logging.Simple
@@ -30,14 +26,20 @@ namespace Common.Logging.Simple
     /// Sends log messages to <see cref="Console.Out" />.
     /// </summary>
     /// <author>Gilles Bayon</author>
-#if PORTABLE
-#else
+    /// <remarks>
+    /// Creates and initializes a logger that writes messages to <see cref="Console.Out" />.
+    /// </remarks>
+    /// <param name="logName">The name, usually type name of the calling class, of the logger.</param>
+    /// <param name="logLevel">The current logging threshold. Messages recieved that are beneath this threshold will not be logged.</param>
+    /// <param name="showLevel">Include the current log level in the log message.</param>
+    /// <param name="showDateTime">Include the current time in the log message.</param>
+    /// <param name="showLogName">Include the instance name in the log message.</param>
+    /// <param name="dateTimeFormat">The date and time format to use in the log message.</param>
     [Serializable]
-#endif
-    public class ConsoleOutLogger : Simple.AbstractSimpleLogger
+    public class ConsoleOutLogger(string logName, LogLevel logLevel, bool showLevel, bool showDateTime, bool showLogName, string dateTimeFormat)
+        : AbstractSimpleLogger(logName, logLevel, showLevel, showDateTime, showLogName, dateTimeFormat)
     {
-#if !SILVERLIGHT
-        private static readonly Dictionary<LogLevel, ConsoleColor> colors = new Dictionary<LogLevel, ConsoleColor>
+        private static readonly Dictionary<LogLevel, ConsoleColor> colors = new()
         {
             { LogLevel.Fatal, ConsoleColor.Red },
             { LogLevel.Error, ConsoleColor.Yellow },
@@ -49,22 +51,6 @@ namespace Common.Logging.Simple
 
         private readonly bool useColor;
 
-#endif
-        /// <summary>
-        /// Creates and initializes a logger that writes messages to <see cref="Console.Out" />.
-        /// </summary>
-        /// <param name="logName">The name, usually type name of the calling class, of the logger.</param>
-        /// <param name="logLevel">The current logging threshold. Messages recieved that are beneath this threshold will not be logged.</param>
-        /// <param name="showLevel">Include the current log level in the log message.</param>
-        /// <param name="showDateTime">Include the current time in the log message.</param>
-        /// <param name="showLogName">Include the instance name in the log message.</param>
-        /// <param name="dateTimeFormat">The date and time format to use in the log message.</param>
-        public ConsoleOutLogger(string logName, LogLevel logLevel, bool showLevel, bool showDateTime, bool showLogName, string dateTimeFormat)
-            : base(logName, logLevel, showLevel, showDateTime, showLogName, dateTimeFormat)
-        {
-        }
-
-#if !SILVERLIGHT
         /// <summary>
         /// Creates and initializes a logger that writes messages to <see cref="Console.Out" />.
         /// </summary>
@@ -81,7 +67,6 @@ namespace Common.Logging.Simple
             this.useColor = useColor;
         }
 
-#endif
         /// <summary>
         /// Do the actual logging by constructing the log message using a <see cref="StringBuilder" /> then
         /// sending the output to <see cref="Console.Out" />.
@@ -92,13 +77,11 @@ namespace Common.Logging.Simple
         protected override void WriteInternal(LogLevel level, object message, Exception e)
         {
             // Use a StringBuilder for better performance
-            StringBuilder sb = new StringBuilder();
+            StringBuilder sb = new();
             FormatOutput(sb, level, message, e);
 
             // Print to the appropriate destination
-#if !SILVERLIGHT
-            ConsoleColor color;
-            if (this.useColor && colors.TryGetValue(level, out color))
+            if (useColor && colors.TryGetValue(level, out ConsoleColor color))
             {
                 var originalColor = Console.ForegroundColor;
                 try
@@ -113,7 +96,6 @@ namespace Common.Logging.Simple
                 }
             }
 
-#endif
             Console.Out.WriteLine(sb.ToString());
         }
     }
