@@ -1,76 +1,56 @@
-﻿using System;
+﻿using Common.Logging;
+using Ipfs.CoreApi;
+using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using Ipfs.CoreApi;
-using Common.Logging;
-using System.Linq;
-using System.Runtime.Serialization;
 
 namespace Ipfs.Engine.CoreApi
 {
-    [DataContract]
-    class DataBlock : IDataBlock
+    internal class BlockApi(IpfsEngine ipfs) : IBlockApi
     {
-        [DataMember]
-        public byte[] DataBytes { get; set; }
-
-        public Stream DataStream { get { return new MemoryStream(DataBytes, false); } }
-
-        [DataMember]
-        public Cid Id  { get; set; }
-
-        [DataMember]
-        public long Size { get; set; }
-    }
-
-    class BlockApi : IBlockApi
-    {
-        static ILog log = LogManager.GetLogger(typeof(BlockApi));
-        static DataBlock emptyDirectory = new DataBlock
+        private static readonly DataBlock emptyDirectory = new()
         {
             DataBytes = ObjectApi.EmptyDirectory.ToArray(),
             Id = ObjectApi.EmptyDirectory.Id,
             Size = ObjectApi.EmptyDirectory.ToArray().Length
         };
-        static DataBlock emptyNode = new DataBlock
+
+        private static readonly DataBlock emptyNode = new()
         {
             DataBytes = ObjectApi.EmptyNode.ToArray(),
             Id = ObjectApi.EmptyNode.Id,
             Size = ObjectApi.EmptyNode.ToArray().Length
         };
 
-        IpfsEngine ipfs;
-        FileStore<Cid, DataBlock> store;
-
-        public BlockApi(IpfsEngine ipfs)
-        {
-            this.ipfs = ipfs;
-        }
+        private static readonly ILog log = LogManager.GetLogger<BlockApi>();
+        private FileStore<Cid, DataBlock> store;
 
         public FileStore<Cid, DataBlock> Store
         {
             get
             {
-                if (store == null)
+                if (store is null)
                 {
-                    var folder = Path.Combine(ipfs.Options.Repository.Folder, "blocks");
+                    string folder = Path.Combine(ipfs.Options.Repository.Folder, "blocks");
                     if (!Directory.Exists(folder))
-                        Directory.CreateDirectory(folder);
+                    {
+                        _ = Directory.CreateDirectory(folder);
+                    }
+
                     store = new FileStore<Cid, DataBlock>
                     {
                         Folder = folder,
                         NameToKey = (cid) => cid.Hash.ToBase32(),
                         KeyToName = (key) => new MultiHash(key.FromBase32()),
-                        Serialize = async (stream, cid, block, cancel) => 
+                        Serialize = async (stream, cid, block, cancel) =>
                         {
-                            await stream.WriteAsync(block.DataBytes, 0, block.DataBytes.Length, cancel).ConfigureAwait(false);
+                            await stream.WriteAsync(block.DataBytes.AsMemory(0, block.DataBytes.Length), cancel).ConfigureAwait(false);
                         },
                         Deserialize = async (stream, cid, cancel) =>
                         {
-                            var block = new DataBlock
+                            DataBlock block = new()
                             {
                                 Id = cid,
                                 Size = stream.Length
@@ -78,8 +58,9 @@ namespace Ipfs.Engine.CoreApi
                             block.DataBytes = new byte[block.Size];
                             for (int i = 0, n; i < block.Size; i += n)
                             {
-                                n = await stream.ReadAsync(block.DataBytes, i, (int)block.Size - i, cancel).ConfigureAwait(false);
+                                n = await stream.ReadAsync(block.DataBytes.AsMemory(i, (int)block.Size - i), cancel).ConfigureAwait(false);
                             }
+
                             return block;
                         }
                     };
@@ -88,14 +69,18 @@ namespace Ipfs.Engine.CoreApi
             }
         }
 
-
-        public async Task<IDataBlock> GetAsync(Cid id, CancellationToken cancel = default(CancellationToken))
+        public async Task<IDataBlock> GetAsync(Cid id, CancellationToken cancel = default)
         {
             // Hack for empty object and empty directory object
             if (id == emptyDirectory.Id)
+            {
                 return emptyDirectory;
+            }
+
             if (id == emptyNode.Id)
+            {
                 return emptyNode;
+            }
 
             // If identity hash, then CID has the content.
             if (id.Hash.IsIdentityHash)
@@ -109,50 +94,30 @@ namespace Ipfs.Engine.CoreApi
             }
 
             // Check the local filesystem for the block.
-            var block = await Store.TryGetAsync(id, cancel).ConfigureAwait(false);
+            DataBlock block = await Store.TryGetAsync(id, cancel).ConfigureAwait(false);
             if (block != null)
             {
                 return block;
             }
 
-            // Query the network, via DHT, for peers that can provide the
-            // content.  As a provider peer is found, it is connected to and
-            // the bitswap want lists are exchanged.  Hopefully the provider will
-            // then send the block to us via bitswap and the get task will finish.
-            using (var queryCancel = CancellationTokenSource.CreateLinkedTokenSource(cancel))
-            {
-                var bitswapGet = ipfs.Bitswap.GetAsync(id, queryCancel.Token).ConfigureAwait(false);
-                var dht = await ipfs.DhtService;
-                var _ = dht.FindProvidersAsync(
-                    id: id,
-                    limit: 20, // TODO: remove this
-                    cancel: queryCancel.Token,
-                    action: (peer) => { var __ = ProviderFoundAsync(peer, queryCancel.Token).ConfigureAwait(false); }
-                );
+            // Query the network, via DHT, for peers that can provide the content. As a provider
+            // peer is found, it is connected to and the bitswap want lists are exchanged. Hopefully
+            // the provider will then send the block to us via bitswap and the get task will finish.
+            using CancellationTokenSource queryCancel = CancellationTokenSource.CreateLinkedTokenSource(cancel);
+            System.Runtime.CompilerServices.ConfiguredTaskAwaitable<IDataBlock> bitswapGet = ipfs.Bitswap.GetAsync(id, queryCancel.Token).ConfigureAwait(false);
+            PeerTalk.Routing.Dht1 dht = await ipfs.DhtService;
+            Task<IEnumerable<Peer>> _ = dht.FindProvidersAsync(
+                id: id,
+                limit: 20, // TODO: remove this
+                cancel: queryCancel.Token,
+                action: (peer) => { System.Runtime.CompilerServices.ConfiguredTaskAwaitable __ = ProviderFoundAsync(peer, queryCancel.Token).ConfigureAwait(false); }
+            );
 
-                var got = await bitswapGet;
-                log.Debug("bitswap got the block");
+            IDataBlock got = await bitswapGet;
+            log.Debug("bitswap got the block");
 
-                queryCancel.Cancel(false); // stop the network query.
-                return got;
-            }
-        }
-
-        async Task ProviderFoundAsync(Peer peer, CancellationToken cancel)
-        {
-            if (cancel.IsCancellationRequested)
-                return;
-
-            log.Debug($"Connecting to provider {peer.Id}");
-            var swarm = await ipfs.SwarmService.ConfigureAwait(false);
-            try
-            {
-                await swarm.ConnectAsync(peer, cancel).ConfigureAwait(false);
-            }
-            catch (Exception e)
-            {
-                log.Warn($"Connection to provider {peer.Id} failed, {e.Message}");
-            }
+            await queryCancel.CancelAsync(); // stop the network query.
+            return got;
         }
 
         public async Task<Cid> PutAsync(
@@ -161,11 +126,11 @@ namespace Ipfs.Engine.CoreApi
             string multiHash = MultiHash.DefaultAlgorithmName,
             string encoding = MultiBase.DefaultAlgorithmName,
             bool pin = false,
-            CancellationToken cancel = default(CancellationToken))
+            CancellationToken cancel = default)
         {
             if (data.Length > ipfs.Options.Block.MaxBlockSize)
             {
-                throw new ArgumentOutOfRangeException("data.Length", $"Block length can not exceed { ipfs.Options.Block.MaxBlockSize}.");
+                throw new ArgumentOutOfRangeException("data.Length", $"Block length can not exceed {ipfs.Options.Block.MaxBlockSize}.");
             }
 
             // Small enough for an inline CID?
@@ -178,9 +143,8 @@ namespace Ipfs.Engine.CoreApi
                 };
             }
 
-            // CID V1 encoding defaulting to base32 which is not
-            // the multibase default. 
-            var cid = new Cid
+            // CID V1 encoding defaulting to base32 which is not the multibase default.
+            Cid cid = new()
             {
                 ContentType = contentType,
                 Hash = MultiHash.ComputeHash(data, multiHash)
@@ -189,13 +153,13 @@ namespace Ipfs.Engine.CoreApi
             {
                 cid.Encoding = encoding;
             }
-            var block = new DataBlock
+            DataBlock block = new()
             {
                 DataBytes = data,
                 Id = cid,
                 Size = data.Length
             };
-            if (await Store.ExistsAsync(cid).ConfigureAwait(false))
+            if (await Store.ExistsAsync(cid, cancel).ConfigureAwait(false))
             {
                 log.DebugFormat("Block '{0}' already present", cid);
             }
@@ -210,17 +174,12 @@ namespace Ipfs.Engine.CoreApi
             }
 
             // Inform the Bitswap service.
-            (await ipfs.BitswapService.ConfigureAwait(false)).Found(block);
+            _ = (await ipfs.BitswapService.ConfigureAwait(false)).Found(block);
 
             // To pin or not.
-            if (pin)
-            {
-                await ipfs.Pin.AddAsync(cid, recursive: false, cancel: cancel).ConfigureAwait(false);
-            }
-            else
-            {
-                await ipfs.Pin.RemoveAsync(cid, recursive: false, cancel: cancel).ConfigureAwait(false);
-            }
+            _ = pin
+                ? await ipfs.Pin.AddAsync(cid, recursive: false, cancel: cancel).ConfigureAwait(false)
+                : await ipfs.Pin.RemoveAsync(cid, recursive: false, cancel: cancel).ConfigureAwait(false);
 
             return cid;
         }
@@ -231,16 +190,14 @@ namespace Ipfs.Engine.CoreApi
             string multiHash = MultiHash.DefaultAlgorithmName,
             string encoding = MultiBase.DefaultAlgorithmName,
             bool pin = false,
-            CancellationToken cancel = default(CancellationToken))
+            CancellationToken cancel = default)
         {
-            using (var ms = new MemoryStream())
-            {
-                await data.CopyToAsync(ms).ConfigureAwait(false);
-                return await PutAsync(ms.ToArray(), contentType, multiHash, encoding, pin, cancel).ConfigureAwait(false);
-            }
+            using MemoryStream ms = new();
+            await data.CopyToAsync(ms, cancel).ConfigureAwait(false);
+            return await PutAsync(ms.ToArray(), contentType, multiHash, encoding, pin, cancel).ConfigureAwait(false);
         }
 
-        public async Task<Cid> RemoveAsync(Cid id, bool ignoreNonexistent = false, CancellationToken cancel = default(CancellationToken))
+        public async Task<Cid> RemoveAsync(Cid id, bool ignoreNonexistent = false, CancellationToken cancel = default)
         {
             if (id.Hash.IsIdentityHash)
             {
@@ -249,14 +206,13 @@ namespace Ipfs.Engine.CoreApi
             if (await Store.ExistsAsync(id, cancel).ConfigureAwait(false))
             {
                 await Store.RemoveAsync(id, cancel).ConfigureAwait(false);
-                await ipfs.Pin.RemoveAsync(id, recursive: false, cancel: cancel).ConfigureAwait(false);
+                _ = await ipfs.Pin.RemoveAsync(id, recursive: false, cancel: cancel).ConfigureAwait(false);
                 return id;
             }
-            if (ignoreNonexistent) return null;
-            throw new KeyNotFoundException($"Block '{id.Encode()}' does not exist.");
+            return ignoreNonexistent ? null : throw new KeyNotFoundException($"Block '{id.Encode()}' does not exist.");
         }
 
-        public async Task<IDataBlock> StatAsync(Cid id, CancellationToken cancel = default(CancellationToken))
+        public async Task<IDataBlock> StatAsync(Cid id, CancellationToken cancel = default)
         {
             if (id.Hash.IsIdentityHash)
             {
@@ -264,7 +220,7 @@ namespace Ipfs.Engine.CoreApi
             }
 
             IDataBlock block = null;
-            var length = await Store.LengthAsync(id, cancel).ConfigureAwait(false);
+            long? length = await Store.LengthAsync(id, cancel).ConfigureAwait(false);
             if (length.HasValue)
             {
                 block = new DataBlock
@@ -275,6 +231,25 @@ namespace Ipfs.Engine.CoreApi
             }
 
             return block;
+        }
+
+        private async Task ProviderFoundAsync(Peer peer, CancellationToken cancel)
+        {
+            if (cancel.IsCancellationRequested)
+            {
+                return;
+            }
+
+            log.Debug($"Connecting to provider {peer.Id}");
+            PeerTalk.Swarm swarm = await ipfs.SwarmService.ConfigureAwait(false);
+            try
+            {
+                _ = await swarm.ConnectAsync(peer, cancel).ConfigureAwait(false);
+            }
+            catch (Exception e)
+            {
+                log.Warn($"Connection to provider {peer.Id} failed, {e.Message}");
+            }
         }
     }
 }
