@@ -1,80 +1,71 @@
 ﻿using Common.Logging;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using Newtonsoft.Json;
-using System.IO;
-using System.Threading.Tasks;
-using System.Threading;
 using Ipfs.CoreApi;
 using Ipfs.Engine.CoreApi;
 using Ipfs.Engine.Cryptography;
 using Ipfs.Engine.Migration;
-using PeerTalk;
-using System.Reflection;
-using PeerTalk.Discovery;
-using Nito.AsyncEx;
 using Makaretu.Dns;
-using System.Collections.Concurrent;
-using System.Security;
-using PeerTalk.SecureCommunication;
+using Nito.AsyncEx;
+using PeerTalk;
 using PeerTalk.Cryptography;
+using PeerTalk.Discovery;
+using PeerTalk.SecureCommunication;
+using System;
+using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Reflection;
+using System.Security;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace Ipfs.Engine
 {
     /// <summary>
-    ///   Implements the <see cref="Ipfs.CoreApi.ICoreApi">Core API</see> which makes it possible to create a decentralised and distributed 
-    ///   application without relying on an "IPFS daemon".
+    /// Implements the <see cref="ICoreApi">Core API</see> which makes it possible to
+    /// create a decentralised and distributed application without relying on an "IPFS daemon".
     /// </summary>
     /// <remarks>
-    ///   The engine should be used as a shared object in your program. It is thread safe (re-entrant) and conserves 
-    ///   resources when only one instance is used.
+    /// The engine should be used as a shared object in your program. It is thread safe (re-entrant)
+    /// and conserves resources when only one instance is used.
     /// </remarks>
     public partial class IpfsEngine : ICoreApi, IService, IDisposable
     {
-        static ILog log = LogManager.GetLogger(typeof(IpfsEngine));
+        private static readonly ILog log = LogManager.GetLogger<IpfsEngine>();
 
-        KeyChain keyChain;
-        SecureString passphrase;
-        ConcurrentBag<Func<Task>> stopTasks = new ConcurrentBag<Func<Task>>();
+        private KeyChain keyChain;
+        private readonly SecureString passphrase;
+        private ConcurrentBag<Func<Task>> stopTasks = [];
 
         /// <summary>
-        ///   Creates a new instance of the <see cref="IpfsEngine"/> class
-        ///   with the IPFS_PASS environment variable.
+        /// Creates a new instance of the <see cref="IpfsEngine"/> class with the IPFS_PASS
+        /// environment variable.
         /// </summary>
-        /// <remarks>
-        ///   Th passphrase must be in the IPFS_PASS environment variable.
-        /// </remarks>
+        /// <remarks>Th passphrase must be in the IPFS_PASS environment variable.</remarks>
         public IpfsEngine()
         {
-            var s = Environment.GetEnvironmentVariable("IPFS_PASS");
-            if (s == null)
-                throw new Exception("The IPFS_PASS environement variable is missing.");
-
+            string s = Environment.GetEnvironmentVariable("IPFS_PASS") ?? throw new Exception("The IPFS_PASS environement variable is missing.");
             passphrase = new SecureString();
-            foreach (var c in s)
+            foreach (char c in s)
             {
-                this.passphrase.AppendChar(c);
+                passphrase.AppendChar(c);
             }
+
             Init();
         }
 
         /// <summary>
-        ///   Creates a new instance of the <see cref="IpfsEngine"/> class
-        ///   with the specified passphrase.
+        /// Creates a new instance of the <see cref="IpfsEngine"/> class with the specified passphrase.
         /// </summary>
-        /// <param name="passphrase">
-        ///   The password used to access the keychain.
-        /// </param>
+        /// <param name="passphrase">The password used to access the keychain.</param>
         /// <remarks>
-        ///   A <b>SecureString</b> copy of the passphrase is made so that the array can be 
-        ///   zeroed out after the call.
+        /// A <b>SecureString</b> copy of the passphrase is made so that the array can be zeroed out
+        /// after the call.
         /// </remarks>
         public IpfsEngine(char[] passphrase)
         {
             this.passphrase = new SecureString();
-            foreach (var c in passphrase)
+            foreach (char c in passphrase)
             {
                 this.passphrase.AppendChar(c);
             }
@@ -82,23 +73,18 @@ namespace Ipfs.Engine
         }
 
         /// <summary>
-        ///   Creates a new instance of the <see cref="IpfsEngine"/> class
-        ///   with the specified passphrase.
+        /// Creates a new instance of the <see cref="IpfsEngine"/> class with the specified passphrase.
         /// </summary>
-        /// <param name="passphrase">
-        ///   The password used to access the keychain.
-        /// </param>
-        /// <remarks>
-        ///  A copy of the <paramref name="passphrase"/> is made.
-        /// </remarks>
+        /// <param name="passphrase">The password used to access the keychain.</param>
+        /// <remarks>A copy of the <paramref name="passphrase"/> is made.</remarks>
         public IpfsEngine(SecureString passphrase)
         {
             this.passphrase = passphrase.Copy();
             Init();
         }
 
-        void Init()
-        { 
+        private void Init()
+        {
             // Init the core api inteface.
             Bitswap = new BitswapApi(this);
             Block = new BlockApi(this);
@@ -124,16 +110,16 @@ namespace Ipfs.Engine
             LocalPeer = new AsyncLazy<Peer>(async () =>
             {
                 log.Debug("Building local peer");
-                var keyChain = await KeyChainAsync().ConfigureAwait(false);
+                KeyChain keyChain = await KeyChainAsync().ConfigureAwait(false);
                 log.Debug("Getting key info about self");
-                var self = await keyChain.FindKeyByNameAsync("self").ConfigureAwait(false);
-                var localPeer = new Peer
+                IKey self = await keyChain.FindKeyByNameAsync("self").ConfigureAwait(false);
+                Peer localPeer = new()
                 {
                     Id = self.Id,
                     PublicKey = await keyChain.GetPublicKeyAsync("self").ConfigureAwait(false),
                     ProtocolVersion = "ipfs/0.1.0"
                 };
-                var version = typeof(IpfsEngine).GetTypeInfo().Assembly.GetName().Version;
+                Version version = typeof(IpfsEngine).GetTypeInfo().Assembly.GetName().Version;
                 localPeer.AgentVersion = $"net-ipfs/{version.Major}.{version.Minor}.{version.Revision}";
                 log.Debug("Built local peer");
                 return localPeer;
@@ -143,20 +129,18 @@ namespace Ipfs.Engine
                 log.Debug("Building swarm service");
                 if (Options.Swarm.PrivateNetworkKey == null)
                 {
-                    var path = Path.Combine(Options.Repository.Folder, "swarm.key");
+                    string path = Path.Combine(Options.Repository.Folder, "swarm.key");
                     if (File.Exists(path))
                     {
-                        using (var x = File.OpenText(path))
-                        {
-                            Options.Swarm.PrivateNetworkKey = new PreSharedKey();
-                            Options.Swarm.PrivateNetworkKey.Import(x);
-                        }
+                        using StreamReader x = File.OpenText(path);
+                        Options.Swarm.PrivateNetworkKey = new PreSharedKey();
+                        Options.Swarm.PrivateNetworkKey.Import(x);
                     }
                 }
-                var peer = await LocalPeer.ConfigureAwait(false);
-                var keyChain = await KeyChainAsync().ConfigureAwait(false);
-                var self = await keyChain.GetPrivateKeyAsync("self").ConfigureAwait(false);
-                var swarm = new Swarm
+                Peer peer = await LocalPeer.ConfigureAwait(false);
+                KeyChain keyChain = await KeyChainAsync().ConfigureAwait(false);
+                Org.BouncyCastle.Crypto.AsymmetricKeyParameter self = await keyChain.GetPrivateKeyAsync("self").ConfigureAwait(false);
+                Swarm swarm = new()
                 {
                     LocalPeer = peer,
                     LocalPeerKey = PeerTalk.Cryptography.Key.CreatePrivateKey(self),
@@ -165,7 +149,9 @@ namespace Ipfs.Engine
                         : new Psk1Protector { Key = Options.Swarm.PrivateNetworkKey }
                 };
                 if (Options.Swarm.PrivateNetworkKey != null)
+                {
                     log.Debug($"Private network {Options.Swarm.PrivateNetworkKey.Fingerprint().ToHexString()}");
+                }
 
                 log.Debug("Built swarm service");
                 return swarm;
@@ -173,7 +159,7 @@ namespace Ipfs.Engine
             BitswapService = new AsyncLazy<BlockExchange.Bitswap>(async () =>
             {
                 log.Debug("Building bitswap service");
-                var bitswap = new BlockExchange.Bitswap
+                BlockExchange.Bitswap bitswap = new()
                 {
                     Swarm = await SwarmService.ConfigureAwait(false),
                     BlockService = Block
@@ -184,7 +170,7 @@ namespace Ipfs.Engine
             DhtService = new AsyncLazy<PeerTalk.Routing.Dht1>(async () =>
             {
                 log.Debug("Building DHT service");
-                var dht = new PeerTalk.Routing.Dht1
+                PeerTalk.Routing.Dht1 dht = new()
                 {
                     Swarm = await SwarmService.ConfigureAwait(false)
                 };
@@ -195,7 +181,7 @@ namespace Ipfs.Engine
             PingService = new AsyncLazy<PeerTalk.Protocols.Ping1>(async () =>
             {
                 log.Debug("Building Ping service");
-                var ping = new PeerTalk.Protocols.Ping1
+                PeerTalk.Protocols.Ping1 ping = new()
                 {
                     Swarm = await SwarmService.ConfigureAwait(false)
                 };
@@ -205,7 +191,7 @@ namespace Ipfs.Engine
             PubSubService = new AsyncLazy<PeerTalk.PubSub.NotificationService>(async () =>
             {
                 log.Debug("Building PubSub service");
-                var pubsub = new PeerTalk.PubSub.NotificationService
+                PeerTalk.PubSub.NotificationService pubsub = new()
                 {
                     LocalPeer = await LocalPeer.ConfigureAwait(false)
                 };
@@ -219,160 +205,141 @@ namespace Ipfs.Engine
         }
 
         /// <summary>
-        ///   The configuration options.
+        /// The configuration options.
         /// </summary>
         public IpfsEngineOptions Options { get; set; } = new IpfsEngineOptions();
 
         /// <summary>
-        ///   Manages the version of the repository.
+        /// Manages the version of the repository.
         /// </summary>
         public MigrationManager MigrationManager { get; set; }
 
-        /// <inheritdoc />
+        /// <inheritdoc/>
         public IBitswapApi Bitswap { get; set; }
 
-        /// <inheritdoc />
+        /// <inheritdoc/>
         public IBlockApi Block { get; set; }
 
-        /// <inheritdoc />
+        /// <inheritdoc/>
         public IBlockRepositoryApi BlockRepository { get; set; }
 
-        /// <inheritdoc />
+        /// <inheritdoc/>
         public IBootstrapApi Bootstrap { get; set; }
 
-        /// <inheritdoc />
+        /// <inheritdoc/>
         public IConfigApi Config { get; set; }
 
-        /// <inheritdoc />
+        /// <inheritdoc/>
         public IDagApi Dag { get; set; }
 
-        /// <inheritdoc />
+        /// <inheritdoc/>
         public IDhtApi Dht { get; set; }
 
-        /// <inheritdoc />
+        /// <inheritdoc/>
         public IDnsApi Dns { get; set; }
 
-        /// <inheritdoc />
+        /// <inheritdoc/>
         public IFileSystemApi FileSystem { get; set; }
 
-        /// <inheritdoc />
+        /// <inheritdoc/>
         public IGenericApi Generic { get; set; }
 
-        /// <inheritdoc />
+        /// <inheritdoc/>
         public IKeyApi Key { get; set; }
 
-        /// <inheritdoc />
+        /// <inheritdoc/>
         public INameApi Name { get; set; }
 
-        /// <inheritdoc />
+        /// <inheritdoc/>
         public IObjectApi Object { get; set; }
 
-        /// <inheritdoc />
+        /// <inheritdoc/>
         public IPinApi Pin { get; set; }
 
-        /// <inheritdoc />
+        /// <inheritdoc/>
         public IPubSubApi PubSub { get; set; }
 
-        /// <inheritdoc />
+        /// <inheritdoc/>
         public ISwarmApi Swarm { get; set; }
 
-        /// <inheritdoc />
+        /// <inheritdoc/>
         public IStatsApi Stats { get; set; }
 
         /// <summary>
-        ///   Provides access to the <see cref="KeyChain"/>.
+        /// Provides access to the <see cref="KeyChain"/>.
         /// </summary>
         /// <param name="cancel">
-        ///   Is used to stop the task.  When cancelled, the <see cref="TaskCanceledException"/> is raised.
+        /// Is used to stop the task. When cancelled, the <see cref="TaskCanceledException"/> is raised.
         /// </param>
         /// <returns>
-        ///   A task that represents the asynchronous operation. The task's result is
-        ///   the <see cref="KeyChain"/>.
+        /// A task that represents the asynchronous operation. The task's result is the <see cref="KeyChain"/>.
         /// </returns>
-        public async Task<KeyChain> KeyChainAsync(CancellationToken cancel = default(CancellationToken))
+        public async Task<KeyChain> KeyChainAsync(CancellationToken cancel = default)
         {
             // TODO: this should be a LazyAsync property.
-            if (keyChain == null)
+            if (keyChain is null)
             {
                 lock (this)
                 {
-                    if (keyChain == null)
+                    keyChain ??= new KeyChain(this)
                     {
-                        keyChain = new KeyChain(this)
-                        {
-                            Options = Options.KeyChain
-                        };
-                     }
+                        Options = Options.KeyChain
+                    };
                 }
 
                 await keyChain.SetPassphraseAsync(passphrase, cancel).ConfigureAwait(false);
-                
+
                 // Maybe create "self" key, this is the local peer's id.
-                var self = await keyChain.FindKeyByNameAsync("self", cancel).ConfigureAwait(false);
-                if (self == null)
-                {
-                    self = await keyChain.CreateAsync("self", null, 0, cancel).ConfigureAwait(false);
-                }
+                IKey self = await keyChain.FindKeyByNameAsync("self", cancel).ConfigureAwait(false);
+                _ = self ?? await keyChain.CreateAsync("self", null, 0, cancel).ConfigureAwait(false);
             }
+
             return keyChain;
         }
 
         /// <summary>
-        ///   Provides access to the local peer.
+        /// Provides access to the local peer.
         /// </summary>
         /// <returns>
-        ///   A task that represents the asynchronous operation. The task's result is
-        ///   a <see cref="Ipfs.Peer"/>.
+        /// A task that represents the asynchronous operation. The task's result is a <see cref="Peer"/>.
         /// </returns>
         public AsyncLazy<Peer> LocalPeer { get; private set; }
 
         /// <summary>
-        ///   Resolve an "IPFS path" to a content ID.
+        /// Resolve an "IPFS path" to a content ID.
         /// </summary>
-        /// <param name="path">
-        ///   A IPFS path, such as "Qm...", "Qm.../a/b/c" or "/ipfs/QM..."
-        /// </param>
+        /// <param name="path">A IPFS path, such as "Qm...", "Qm.../a/b/c" or "/ipfs/QM..."</param>
         /// <param name="cancel">
-        ///   Is used to stop the task.  When cancelled, the <see cref="TaskCanceledException"/> is raised.
+        /// Is used to stop the task. When cancelled, the <see cref="TaskCanceledException"/> is raised.
         /// </param>
-        /// <returns>
-        ///   The content ID of <paramref name="path"/>.
-        /// </returns>
-        /// <exception cref="ArgumentException">
-        ///   The <paramref name="path"/> cannot be resolved.
-        /// </exception>
-        public async Task<Cid> ResolveIpfsPathToCidAsync (string path, CancellationToken cancel = default(CancellationToken))
+        /// <returns>The content ID of <paramref name="path"/>.</returns>
+        /// <exception cref="ArgumentException">The <paramref name="path"/> cannot be resolved.</exception>
+        public async Task<Cid> ResolveIpfsPathToCidAsync(string path, CancellationToken cancel = default)
         {
-            var r = await Generic.ResolveAsync(path, true, cancel).ConfigureAwait(false);
-            return Cid.Decode(r.Remove(0, 6));  // strip '/ipfs/'.
+            string r = await Generic.ResolveAsync(path, true, cancel).ConfigureAwait(false);
+            return Cid.Decode(r[6..]);  // strip '/ipfs/'.
         }
 
         /// <summary>
-        ///   Determines if the engine has started.
+        /// Determines if the engine has started.
         /// </summary>
-        /// <value>
-        ///   <b>true</b> if the engine has started; otherwise, <b>false</b>.
-        /// </value>
+        /// <value><b>true</b> if the engine has started; otherwise, <b>false</b>.</value>
         /// <seealso cref="Start"/>
         /// <seealso cref="StartAsync"/>
-        public bool IsStarted => stopTasks.Count > 0;
+        public bool IsStarted => !stopTasks.IsEmpty;
 
         /// <summary>
-        ///   Starts the network services.
+        /// Starts the network services.
         /// </summary>
-        /// <returns>
-        ///   A task that represents the asynchronous operation.
-        /// </returns>
+        /// <returns>A task that represents the asynchronous operation.</returns>
         /// <remarks>
-        ///   Starts the various IPFS and PeerTalk network services.  This should
-        ///   be called after any configuration changes.
+        /// Starts the various IPFS and PeerTalk network services. This should be called after any
+        /// configuration changes.
         /// </remarks>
-        /// <exception cref="Exception">
-        ///   When the engine is already started.
-        /// </exception>
+        /// <exception cref="Exception">When the engine is already started.</exception>
         public async Task StartAsync()
         {
-            if (stopTasks.Count > 0)
+            if (!stopTasks.IsEmpty)
             {
                 throw new Exception("IPFS engine is already started.");
             }
@@ -381,64 +348,58 @@ namespace Ipfs.Engine
             await MigrationManager.MirgrateToVersionAsync(MigrationManager.LatestVersion)
                 .ConfigureAwait(false);
 
-            var localPeer = await LocalPeer.ConfigureAwait(false);
-            log.Debug("starting " + localPeer.Id);
+            Peer localPeer = await LocalPeer.ConfigureAwait(false);
+            log.Debug($"starting {localPeer.Id}");
 
             // Everybody needs the swarm.
-            var swarm = await SwarmService.ConfigureAwait(false);
-            stopTasks.Add(async () =>
-            {
-                await swarm.StopAsync().ConfigureAwait(false);
-            });
+            Swarm swarm = await SwarmService.ConfigureAwait(false);
+            stopTasks.Add(swarm.StopAsync);
             await swarm.StartAsync().ConfigureAwait(false);
 
-            var peerManager = new PeerManager { Swarm = swarm };
+            PeerManager peerManager = new() { Swarm = swarm };
             await peerManager.StartAsync().ConfigureAwait(false);
-            stopTasks.Add(async () =>
-            {
-                await peerManager.StopAsync().ConfigureAwait(false);
-            });
+            stopTasks.Add(peerManager.StopAsync);
 
             // Start the primary services.
-            var tasks = new List<Func<Task>>
-            {
+            List<Func<Task>> tasks =
+            [
                 async () =>
                 {
-                    var bitswap = await BitswapService.ConfigureAwait(false);
+                    BlockExchange.Bitswap bitswap = await BitswapService.ConfigureAwait(false);
                     stopTasks.Add(async () => await bitswap.StopAsync().ConfigureAwait(false));
                     await bitswap.StartAsync().ConfigureAwait(false);
                 },
                 async () =>
                 {
-                    var dht = await DhtService.ConfigureAwait(false);
+                    PeerTalk.Routing.Dht1 dht = await DhtService.ConfigureAwait(false);
                     stopTasks.Add(async () => await dht.StopAsync().ConfigureAwait(false));
                     await dht.StartAsync().ConfigureAwait(false);
                 },
                 async () =>
                 {
-                    var ping = await PingService.ConfigureAwait(false);
+                    PeerTalk.Protocols.Ping1 ping = await PingService.ConfigureAwait(false);
                     stopTasks.Add(async () => await ping.StopAsync().ConfigureAwait(false));
                     await ping.StartAsync().ConfigureAwait(false);
                 },
                 async () =>
                 {
-                    var pubsub = await PubSubService.ConfigureAwait(false);
+                    PeerTalk.PubSub.NotificationService pubsub = await PubSubService.ConfigureAwait(false);
                     stopTasks.Add(async () => await pubsub.StopAsync().ConfigureAwait(false));
                     await pubsub.StartAsync().ConfigureAwait(false);
                 },
-            };
+            ];
 
             log.Debug("waiting for services to start");
             await Task.WhenAll(tasks.Select(t => t())).ConfigureAwait(false);
 
             // Starting listening to the swarm.
-            var json = await Config.GetAsync("Addresses.Swarm").ConfigureAwait(false);
-            var numberListeners = 0;
-            foreach (string a in json)
+            Newtonsoft.Json.Linq.JToken json = await Config.GetAsync("Addresses.Swarm").ConfigureAwait(false);
+            int numberListeners = 0;
+            foreach (string a in json.Select(v => (string)v))
             {
                 try
                 {
-                    await swarm.StartListeningAsync(a).ConfigureAwait(false);
+                    _ = await swarm.StartListeningAsync(a).ConfigureAwait(false);
                     ++numberListeners;
                 }
                 catch (Exception e)
@@ -447,38 +408,33 @@ namespace Ipfs.Engine
                     // eat the exception
                 }
             }
+
             if (numberListeners == 0)
             {
                 log.Error("No listeners were created.");
             }
 
-            // Now that the listener addresses are established, the discovery 
-            // services can begin.
+            // Now that the listener addresses are established, the discovery services can begin.
             MulticastService multicast = null;
             if (!Options.Discovery.DisableMdns)
             {
                 multicast = new MulticastService();
-#pragma warning disable CS1998
-                stopTasks.Add(async () => multicast.Dispose());
-#pragma warning restore CS1998
+                stopTasks.Add(() => Task.Run(multicast.Dispose));
             }
 
-            var autodialer = new AutoDialer(swarm)
+            AutoDialer autodialer = new(swarm)
             {
                 MinConnections = Options.Swarm.MinConnections
             };
-#pragma warning disable CS1998
-            stopTasks.Add(async () => autodialer.Dispose());
-#pragma warning restore CS1998
+            stopTasks.Add(() => Task.Run(autodialer.Dispose));
 
-            tasks = new List<Func<Task>>
-            {
+            tasks =
+            [
                 // Bootstrap discovery
                 async () =>
                 {
-                    var bootstrap = new PeerTalk.Discovery.Bootstrap
-                    {
-                        Addresses = await this.Bootstrap.ListAsync()
+                    Bootstrap bootstrap = new() {
+                        Addresses = await Bootstrap.ListAsync()
                     };
                     bootstrap.PeerDiscovered += OnPeerDiscovered;
                     stopTasks.Add(async () => await bootstrap.StopAsync().ConfigureAwait(false));
@@ -487,10 +443,7 @@ namespace Ipfs.Engine
                 // New multicast DNS discovery
                 async () =>
                 {
-                    if (Options.Discovery.DisableMdns)
-                        return;
-                    var mdns = new PeerTalk.Discovery.MdnsNext
-                    {
+                    if (Options.Discovery.DisableMdns) { return; } MdnsNext mdns = new() {
                         LocalPeer = localPeer,
                         MulticastService = multicast
                     };
@@ -505,10 +458,7 @@ namespace Ipfs.Engine
                 // Old style JS multicast DNS discovery
                 async () =>
                 {
-                    if (Options.Discovery.DisableMdns || Options.Swarm.PrivateNetworkKey != null)
-                        return;
-                    var mdns = new PeerTalk.Discovery.MdnsJs
-                    {
+                    if (Options.Discovery.DisableMdns || Options.Swarm.PrivateNetworkKey != null) { return; } MdnsJs mdns = new() {
                         LocalPeer = localPeer,
                         MulticastService = multicast
                     };
@@ -519,10 +469,7 @@ namespace Ipfs.Engine
                 // Old style GO multicast DNS discovery
                 async () =>
                 {
-                    if (Options.Discovery.DisableMdns || Options.Swarm.PrivateNetworkKey != null)
-                        return;
-                    var mdns = new PeerTalk.Discovery.MdnsGo
-                    {
+                    if (Options.Discovery.DisableMdns || Options.Swarm.PrivateNetworkKey != null) { return; } MdnsGo mdns = new() {
                         LocalPeer = localPeer,
                         MulticastService = multicast
                     };
@@ -530,15 +477,13 @@ namespace Ipfs.Engine
                     stopTasks.Add(async () => await mdns.StopAsync().ConfigureAwait(false));
                     await mdns.StartAsync().ConfigureAwait(false);
                 },
-                async () => 
+                async () =>
                 {
-                    if (Options.Discovery.DisableRandomWalk)
-                        return;
-                    var randomWalk = new RandomWalk { Dht = Dht };
+                    if (Options.Discovery.DisableRandomWalk) { return; } RandomWalk randomWalk = new() { Dht = Dht };
                     stopTasks.Add(async () => await randomWalk.StopAsync().ConfigureAwait(false));
                     await randomWalk.StartAsync().ConfigureAwait(false);
                 }
-            };
+            ];
             log.Debug("waiting for discovery services to start");
             await Task.WhenAll(tasks.Select(t => t())).ConfigureAwait(false);
 
@@ -548,21 +493,17 @@ namespace Ipfs.Engine
         }
 
         /// <summary>
-        ///   Stops the running services.
+        /// Stops the running services.
         /// </summary>
-        /// <returns>
-        ///   A task that represents the asynchronous operation.
-        /// </returns>
-        /// <remarks>
-        ///   Multiple calls are okay.
-        /// </remarks>
+        /// <returns>A task that represents the asynchronous operation.</returns>
+        /// <remarks>Multiple calls are okay.</remarks>
         public async Task StopAsync()
         {
             log.Debug("stopping");
             try
             {
-                var tasks = stopTasks.ToArray();
-                stopTasks = new ConcurrentBag<Func<Task>>();
+                Func<Task>[] tasks = stopTasks.ToArray();
+                stopTasks = [];
                 await Task.WhenAll(tasks.Select(t => t())).ConfigureAwait(false);
             }
             catch (Exception e)
@@ -570,8 +511,8 @@ namespace Ipfs.Engine
                 log.Error("Failure when stopping the engine", e);
             }
 
-            // Many services use cancellation to stop.  A cancellation may not run
-            // immediately, so we need to give them some.
+            // Many services use cancellation to stop. A cancellation may not run immediately, so we
+            // need to give them some.
             // TODO: Would be nice to make this deterministic.
             await Task.Delay(TimeSpan.FromMilliseconds(100)).ConfigureAwait(false);
 
@@ -579,32 +520,28 @@ namespace Ipfs.Engine
         }
 
         /// <summary>
-        ///   A synchronous start.
+        /// A synchronous start.
         /// </summary>
-        /// <remarks>
-        ///   Calls <see cref="StartAsync"/> and waits for it to complete.
-        /// </remarks>
+        /// <remarks>Calls <see cref="StartAsync"/> and waits for it to complete.</remarks>
         public void Start()
         {
-            StartAsync().ConfigureAwait(false).GetAwaiter().GetResult();
+            StartAsync().RunSynchronously();
         }
 
         /// <summary>
-        ///   A synchronous stop.
+        /// A synchronous stop.
         /// </summary>
-        /// <remarks>
-        ///   Calls <see cref="StopAsync"/> and waits for it to complete.
-        /// </remarks>
+        /// <remarks>Calls <see cref="StopAsync"/> and waits for it to complete.</remarks>
         public void Stop()
         {
             log.Debug("stopping");
             try
             {
-                var tasks = stopTasks.ToArray();
-                stopTasks = new ConcurrentBag<Func<Task>>();
-                foreach (var task in tasks)
+                Func<Task>[] tasks = stopTasks.ToArray();
+                stopTasks = [];
+                foreach (Func<Task> task in tasks)
                 {
-                    task().ConfigureAwait(false).GetAwaiter().GetResult();
+                    task().RunSynchronously();
                 }
             }
             catch (Exception e)
@@ -614,46 +551,43 @@ namespace Ipfs.Engine
         }
 
         /// <summary>
-        ///   Manages communication with other peers.
+        /// Manages communication with other peers.
         /// </summary>
         public AsyncLazy<Swarm> SwarmService { get; private set; }
 
         /// <summary>
-        ///   Manages publishng and subscribing to messages.
+        /// Manages publishng and subscribing to messages.
         /// </summary>
         public AsyncLazy<PeerTalk.PubSub.NotificationService> PubSubService { get; private set; }
 
         /// <summary>
-        ///   Exchange blocks with other peers.
+        /// Exchange blocks with other peers.
         /// </summary>
         public AsyncLazy<BlockExchange.Bitswap> BitswapService { get; private set; }
 
         /// <summary>
-        ///   Finds information with a distributed hash table.
+        /// Finds information with a distributed hash table.
         /// </summary>
         public AsyncLazy<PeerTalk.Routing.Dht1> DhtService { get; private set; }
 
         /// <summary>
-        ///   Determines latency to a peer.
+        /// Determines latency to a peer.
         /// </summary>
         public AsyncLazy<PeerTalk.Protocols.Ping1> PingService { get; private set; }
 
-        #pragma warning disable VSTHRD100 // Avoid async void methods
         /// <summary>
-        ///   Fired when a peer is discovered.
+        /// Fired when a peer is discovered.
         /// </summary>
         /// <param name="sender"></param>
         /// <param name="peer"></param>
-        /// <remarks>
-        ///   Registers the peer with the <see cref="SwarmService"/>.
-        /// </remarks>
-        async void OnPeerDiscovered(object sender, Peer peer)
-#pragma warning restore VSTHRD100 // Avoid async void methods
+        /// <remarks>Registers the peer with the <see cref="SwarmService"/>.</remarks>
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Usage", "VSTHRD100:Avoid async void methods", Justification = "<Pending>")]
+        private async void OnPeerDiscovered(object sender, Peer peer)
         {
             try
             {
-                var swarm = await SwarmService.ConfigureAwait(false);
-                swarm.RegisterPeer(peer);
+                Swarm swarm = await SwarmService.ConfigureAwait(false);
+                _ = swarm.RegisterPeer(peer);
             }
             catch (Exception ex)
             {
@@ -662,37 +596,37 @@ namespace Ipfs.Engine
             }
         }
 
-        #region IDisposable Support
-        bool disposedValue = false; // To detect redundant calls
+        private bool disposed = false; // To detect redundant calls
 
         /// <summary>
-        ///  Releases the unmanaged and optionally managed resources.
+        /// Releases the unmanaged and optionally managed resources.
         /// </summary>
         /// <param name="disposing">
-        ///   <b>true</b> to release both managed and unmanaged resources; <b>false</b> 
-        ///   to release only unmanaged resources.
+        /// <b>true</b> to release both managed and unmanaged resources; <b>false</b> to release
+        /// only unmanaged resources.
         /// </param>
         protected virtual void Dispose(bool disposing)
         {
-            if (!disposedValue)
+            if (disposed)
             {
-                disposedValue = true;
+                return;
+            }
+            disposed = true;
 
-                if (disposing)
-                {
-                    passphrase?.Dispose();
-                    Stop();
-                }
+            if (disposing)
+            {
+                passphrase?.Dispose();
+                Stop();
             }
         }
 
         /// <summary>
-        ///   Performs application-defined tasks associated with freeing, releasing, or resetting unmanaged resources.
+        /// Performs application-defined tasks associated with freeing, releasing, or resetting
+        /// unmanaged resources.
         /// </summary>
         public void Dispose()
         {
             Dispose(true);
         }
-        #endregion
     }
 }
