@@ -1,31 +1,18 @@
-using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
 using Ipfs.CoreApi;
-using Ipfs.Engine;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Newtonsoft.Json.Serialization;
-using Swashbuckle.AspNetCore.Swagger;
 
 namespace Ipfs.Server
 {
     /// <summary>
     ///   Startup steps.
     /// </summary>
-    class Startup
+    class Startup(IConfiguration configuration)
     {
-
-        public Startup(IConfiguration configuration)
-        {
-            Configuration = configuration;
-        }
-
-        public IConfiguration Configuration { get; }
+        public IConfiguration Configuration { get; } = configuration;
 
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
@@ -35,19 +22,18 @@ namespace Ipfs.Server
             services.AddMvc()
                 .AddJsonOptions(jo =>
                 {
-                    jo.SerializerSettings.ContractResolver = new DefaultContractResolver()
-                    {
-                        NamingStrategy = new DefaultNamingStrategy()
-                    };
-                })
-                ;
+                    //jo.SerializerSettings.ContractResolver = new DefaultContractResolver()
+                    //{
+                    //    NamingStrategy = new DefaultNamingStrategy()
+                    //};
+                });
 
             // Register the Swagger generator, defining 1 or more Swagger documents
             services.AddSwaggerGen(c =>
             {
-                c.SwaggerDoc("v0", new Info {
+                c.SwaggerDoc("v0", new Microsoft.OpenApi.Models.OpenApiInfo {
                     Title = "IPFS HTTP API",
-                    Description = "The API for interacting with IPFS nodes.",  
+                    Description = "The API for interacting with IPFS nodes.",
                     Version = "v0" });
 
                 var path = System.Reflection.Assembly.GetExecutingAssembly().Location;
@@ -78,7 +64,7 @@ namespace Ipfs.Server
             // Enable middleware to serve generated Swagger as a JSON endpoint.
             app.UseSwagger();
 
-            // Enable middleware to serve swagger-ui (HTML, JS, CSS, etc.), 
+            // Enable middleware to serve swagger-ui (HTML, JS, CSS, etc.),
             // specifying the Swagger JSON endpoint.
             app.UseSwaggerUI(c =>
             {
